updated index.js to reflect new changes to file upload
